Refactor integration tests
